+++ examples/1-replace/arrows.ts
@@ -1,0 +1,6 @@
+type N = number;
+
+const sq = (n: N) => n * n;
+const up = (n: N) => n + 1;
+
+export const y = up(sq(2));