+++ examples/1-replace/lambda.el
@@ -1,2 +1,1 @@
 (lambda (x) (>= x 0))
-(lambda (y) (sqrt y))